reorg a bunch of types
