+++ src/fs/metadata.ts
@@ -1,5 +1,12 @@
-import { Metadata } from './types'
 import * as semver from './semver'
+import { SemVer } from './semver'
+
+export type Metadata = {
+  isFile: boolean
+  mtime: number
+  ctime: number
+  version: SemVer
+}
 
 export const empty = (): Metadata => ({
   isFile: false,
